refs #5626. Fix warning

+++ Code/Mantid/Framework/Algorithms/src/NormaliseByDetector.cpp
@@ -265,7 +265,7 @@ namespace Mantid
     MatrixWorkspace_sptr NormaliseByDetector::processHistograms(MatrixWorkspace_sptr inWS)
     {
       const size_t nHistograms = inWS->getNumberHistograms();
-      const size_t progress_items = nHistograms * 1.2;
+      const size_t progress_items = static_cast<size_t>(double(nHistograms) * 1.2);
       Progress prog(this,0.0,1.0, progress_items);
       // Clone the input workspace to create a template for the denominator workspace.
       IAlgorithm_sptr cloneAlg = this->createSubAlgorithm("CloneWorkspace", 0.0, 0.1, true);
